feat: add Tailwind sort feature and refactor core
- Remove package-lock.json.
- Bump version to 0.3.2 and update the test script to run from src/test.
- Add test execution steps to the README documentation.
- Introduce default configuration constants for dynamic class functions, ignore comment, and custom class regex.
- Add a central index that re-exports document utilities, patterns, parsing helpers, selection functions, and related types.
- Move the document parsing utility into the core module.
- Add CLASS_ATTRIBUTE_REGEX constant to match class, className, ngClass, and class:list attribute values in various quoting styles.
- Refactor the selections module to use offset‑based keys, centralize regex creation with flag handling and caching, add compiled custom‑regex and dynamic call regex utilities, and streamline class attribute and custom regex processing for improved performance.
- Add type definitions for custom class regex settings and class selection options.
- Make the Tailwind class cleaner use regex token parsing, conditionally create duplicate and conflict sets based on flags, and return the original class list when no cleaning is requested.
- Add memoization cache for class conflict keys to reduce redundant conflict‑key calculations.
- Add Tailwind core index that re‑exports cleanClassList and getClassConflictKey.
- Refactor test suite by moving individual tests into separate modules and importing them.
- Refactor extension to import sorting utilities from a unified feature module.
- Add command to sort Tailwind classes in the active editor using configurable options and clean duplicates and conflicts.
- Refactor the Tailwind class sorting command into a feature module with a new `sortEditorClasses` function that receives the editor and selection options directly, updating imports and removing internal configuration handling.
- Add an index module that exports `sortActiveEditorClasses` and `registerSortOnSave`.
- Move the sort‑on‑save listener into the sort‑tailwind‑classes feature.
- Add unit tests for document parsing utilities to verify regex escaping, parenthesis matching while skipping string contents, and quoted string range extraction that ignores interpolated template literals.
- Add comprehensive tests for class selection parsing, covering ignore markers, multiline class attributes, custom and tuple regex patterns, and regex caching.
- Add mock document creation and selection text extraction helpers for testing.
- Add a mock VS Code module that provides Selection and Position classes with a compareTo method for testing.
- Add comprehensive tests for cleanClassList covering duplicate removal, conflict resolution, combined cleanup, non‑conflicting spacing, and whitespace normalization.
- Add tests verifying Tailwind class conflict key generation across variants, utility groups, and non‑conflicting utilities.
- Delete the constants module that defined the class attribute regex, default dynamic class functions, ignore comment marker, custom class regex settings, and associated type definitions.
- Delete the document utility module and its exported functions.
- Delete the Tailwind utility re-export file that previously exported getClassConflictKey and cleanClassList.
- Exclude the src/test directory from TypeScript compilation.
- Exclude the test directory pattern from the release TypeScript configuration.
