+++ src/config/defaults.ts
@@ -1,0 +1,5 @@
+import type { CustomClassRegexSetting } from "../core/document";
+
+export const DEFAULT_DYNAMIC_CLASS_FUNCTIONS = ["clsx", "cn", "classnames"];
+export const DEFAULT_IGNORE_COMMENT = "tailwindocd-ignore";
+export const DEFAULT_CUSTOM_CLASS_REGEX: CustomClassRegexSetting[] = [];
