+++ src/core/document/patterns.ts
@@ -1,0 +1,2 @@
+export const CLASS_ATTRIBUTE_REGEX =
+  /\b(?:class|className|ngClass|class:list)\s*=\s*(?:"([^"]*)"|'([^']*)'|`([^`]*)`)/g;
