+++ src/core/document/types.ts
@@ -1,0 +1,7 @@
+export type CustomClassRegexSetting = string | [string] | [string, string];
+
+export type ClassSelectionOptions = {
+  dynamicClassFunctions: string[];
+  ignoreCommentMarker: string;
+  customClassRegex: CustomClassRegexSetting[];
+};
