+++ src/core/tailwind/index.ts
@@ -1,0 +1,2 @@
+export { cleanClassList } from "./cleaner";
+export { getClassConflictKey } from "./conflicts";
