+++ src/features/sort-tailwind-classes/index.ts
@@ -1,0 +1,2 @@
+export { sortActiveEditorClasses } from "./command";
+export { registerSortOnSave } from "./on-save";
