--- src/util/document.ts
@@ -1,6 +1,0 @@
-export {
-  escapeForRegex,
-  findClosingParen,
-  getQuotedStringRanges,
-} from "./document/parsing";
-export { getClassSelections, shouldIgnoreLine } from "./document/selections";
