--- src/util/document/selections.ts
@@ -1,229 +1,0 @@
-import * as vscode from "vscode";
-import {
-  CLASS_ATTRIBUTE_REGEX,
-  ClassSelectionOptions,
-  CustomClassRegexSetting,
-} from "../constants";
-import {
-  escapeForRegex,
-  findClosingParen,
-  getQuotedStringRanges,
-} from "./parsing";
-
-function buildSelectionKey(start: vscode.Position, end: vscode.Position): string {
-  return `${start.line}:${start.character}-${end.line}:${end.character}`;
-}
-
-export function shouldIgnoreLine(
-  document: vscode.TextDocument,
-  lineNumber: number,
-  ignoreCommentMarker: string,
-): boolean {
-  if (!ignoreCommentMarker.trim()) return false;
-  const lineText = document.lineAt(lineNumber).text;
-  return lineText.includes(ignoreCommentMarker);
-}
-
-function shouldIgnoreSelection(
-  document: vscode.TextDocument,
-  start: vscode.Position,
-  end: vscode.Position,
-  ignoreCommentMarker: string,
-): boolean {
-  for (let line = start.line; line <= end.line; line += 1) {
-    if (shouldIgnoreLine(document, line, ignoreCommentMarker)) {
-      return true;
-    }
-  }
-
-  return false;
-}
-
-function ensureGlobalRegex(pattern: RegExp): RegExp {
-  const flags = pattern.flags.includes("g")
-    ? pattern.flags
-    : `${pattern.flags}g`;
-  return new RegExp(pattern.source, flags);
-}
-
-function ensureGlobalAndIndicesRegex(pattern: RegExp): RegExp {
-  const globalPattern = ensureGlobalRegex(pattern);
-  const flags = globalPattern.flags.includes("d")
-    ? globalPattern.flags
-    : `${globalPattern.flags}d`;
-  return new RegExp(globalPattern.source, flags);
-}
-
-function parseRegexPattern(patternSource: string): RegExp | null {
-  const trimmedPattern = patternSource.trim();
-  if (!trimmedPattern) return null;
-
-  try {
-    if (trimmedPattern.startsWith("/")) {
-      const lastSlashIndex = trimmedPattern.lastIndexOf("/");
-      if (lastSlashIndex > 0) {
-        const source = trimmedPattern.slice(1, lastSlashIndex);
-        const flags = trimmedPattern.slice(lastSlashIndex + 1);
-        return ensureGlobalRegex(new RegExp(source, flags));
-      }
-    }
-
-    return new RegExp(trimmedPattern, "g");
-  } catch {
-    return null;
-  }
-}
-
-type IndexedMatch = RegExpExecArray & {
-  indices?: Array<[number, number] | undefined>;
-};
-
-function getFirstCaptureRange(match: IndexedMatch): [number, number] | null {
-  const captureIndices = match.indices?.[1];
-  if (!captureIndices) return null;
-
-  const capturedValue = match[1];
-  if (capturedValue === undefined) return null;
-
-  return captureIndices;
-}
-
-function addCustomRegexSelections(
-  text: string,
-  customClassRegex: CustomClassRegexSetting[],
-  addSelection: (startOffset: number, endOffset: number) => void,
-): void {
-  for (const entry of customClassRegex) {
-    const patterns = Array.isArray(entry) ? entry : [entry];
-    const outerPattern = patterns[0];
-    const innerPattern = patterns[1] ?? null;
-
-    if (!outerPattern) continue;
-
-    const outerRegex = parseRegexPattern(outerPattern);
-    if (!outerRegex) continue;
-
-    const indexedOuterRegex = ensureGlobalAndIndicesRegex(outerRegex);
-
-    let outerMatch: IndexedMatch | null;
-    while ((outerMatch = indexedOuterRegex.exec(text) as IndexedMatch | null) !== null) {
-      if (!outerMatch[0]) {
-        indexedOuterRegex.lastIndex += 1;
-        continue;
-      }
-
-      const containerRange = getFirstCaptureRange(outerMatch);
-      if (!containerRange) continue;
-
-      const [outerMatchStart, outerMatchEnd] = containerRange;
-      const containerText = text.slice(outerMatchStart, outerMatchEnd);
-
-      if (!innerPattern) {
-        addSelection(outerMatchStart, outerMatchEnd);
-        continue;
-      }
-
-      const innerRegex = parseRegexPattern(innerPattern);
-      if (!innerRegex) continue;
-
-      const indexedInnerRegex = ensureGlobalAndIndicesRegex(innerRegex);
-
-      let innerMatch: IndexedMatch | null;
-      while (
-        (innerMatch = indexedInnerRegex.exec(containerText) as IndexedMatch | null) !== null
-      ) {
-        if (!innerMatch[0]) {
-          indexedInnerRegex.lastIndex += 1;
-          continue;
-        }
-
-        const innerRange = getFirstCaptureRange(innerMatch);
-        if (!innerRange) continue;
-
-        const [innerMatchStart, innerMatchEnd] = innerRange;
-        addSelection(
-          outerMatchStart + innerMatchStart,
-          outerMatchStart + innerMatchEnd,
-        );
-      }
-    }
-  }
-}
-
-export function getClassSelections(
-  document: vscode.TextDocument,
-  options: ClassSelectionOptions,
-): vscode.Selection[] {
-  const text = document.getText();
-  const selections: vscode.Selection[] = [];
-  const seenKeys = new Set<string>();
-
-  function addSelection(startOffset: number, endOffset: number): void {
-    if (endOffset <= startOffset) return;
-
-    const start = document.positionAt(startOffset);
-    const end = document.positionAt(endOffset);
-    if (
-      shouldIgnoreSelection(
-        document,
-        start,
-        end,
-        options.ignoreCommentMarker,
-      )
-    ) {
-      return;
-    }
-
-    const selectionKey = buildSelectionKey(start, end);
-    if (seenKeys.has(selectionKey)) return;
-
-    seenKeys.add(selectionKey);
-    selections.push(new vscode.Selection(start, end));
-  }
-
-  let match: RegExpExecArray | null;
-
-  while ((match = CLASS_ATTRIBUTE_REGEX.exec(text)) !== null) {
-    const classValue = match[1] ?? match[2] ?? match[3];
-    if (!classValue || !classValue.trim()) continue;
-
-    const classValueOffset = match.index + match[0].indexOf(classValue);
-    addSelection(classValueOffset, classValueOffset + classValue.length);
-  }
-
-  const escapedDynamicFunctionNames = options.dynamicClassFunctions
-    .map((name) => name.trim())
-    .filter(Boolean)
-    .map(escapeForRegex);
-
-  if (escapedDynamicFunctionNames.length > 0) {
-    const dynamicCallRegex = new RegExp(
-      `\\b(?:${escapedDynamicFunctionNames.join("|")})\\s*\\(`,
-      "g",
-    );
-
-    while ((match = dynamicCallRegex.exec(text)) !== null) {
-      const openParenIndex = match.index + match[0].lastIndexOf("(");
-      const closeParenIndex = findClosingParen(text, openParenIndex);
-      if (closeParenIndex < 0) continue;
-
-      const argsStart = openParenIndex + 1;
-      const argsText = text.slice(argsStart, closeParenIndex);
-      const quotedRanges = getQuotedStringRanges(argsText, argsStart);
-
-      for (const quotedRange of quotedRanges) {
-        const classValue = text.slice(quotedRange.start, quotedRange.end);
-        if (!classValue.trim()) continue;
-
-        addSelection(quotedRange.start, quotedRange.end);
-      }
-    }
-  }
-
-  if (options.customClassRegex.length > 0) {
-    addCustomRegexSelections(text, options.customClassRegex, addSelection);
-  }
-
-  selections.sort((a, b) => a.start.compareTo(b.start));
-  return selections;
-}
