--- src/util/tailwind.ts
@@ -1,2 +1,0 @@
-export { getClassConflictKey } from "./tailwind/conflicts";
-export { cleanClassList } from "./tailwind/cleaner";
