+++ tsconfig.json
@@ -20,6 +20,7 @@
     "src/**/*.ts"
   ],
   "exclude": [
+    "src/test/**",
     "src/**/*.test.ts",
     "src/**/*.spec.ts"
   ]
